* commit 'fac08e956ea87c1a59d2d110a73feb4b08469f45':
  fix lint
  fix test
  fix test
  fix lint and circular dependency issue
  fix after merge and review
  add one more test
  fix all the tests
  fix test temp
  fix test and add auth test
  fix test
  async visitor support
  fix test
  fix tests
  add more validation
  Add test for timeslot item service
  timeslot item delete verification
  timeslot item

+++ src/components/timeslotItems/timeslotItems.service.ts
@@ -6,6 +6,9 @@ import { ErrorCodeV2, MOLErrorV2 } from 'mol-lib-api-contract';
 import { TimeslotItemsRepository } from './timeslotItems.repository';
 import { TimeOfDay, TimeslotItem, TimeslotsSchedule } from '../../models';
 import { DeleteResult } from 'typeorm';
+import { TimeslotItemsActionAuthVisitor } from './timeslotItems.auth';
+import { UserContext } from '../../infrastructure/auth/userContext';
+import { ServicesRepository } from '../services/services.repository';
 
 @InRequestScope
 export class TimeslotItemsService {
@@ -13,6 +16,32 @@ export class TimeslotItemsService {
 	private timeslotsScheduleRepository: TimeslotsScheduleRepository;
 	@Inject
 	private timeslotItemsRepository: TimeslotItemsRepository;
+	@Inject
+	private servicesRepository: ServicesRepository;
+	@Inject
+	private userContext: UserContext;
+
+	private async verifyActionPermission(timeslotSchedule: TimeslotsSchedule): Promise<void> {
+		const authGroups = await this.userContext.getAuthGroups();
+		let timeslotScheduleData = timeslotSchedule;
+		if (!timeslotSchedule._service && !timeslotSchedule._serviceProvider) {
+			timeslotScheduleData = await this.timeslotsScheduleRepository.getTimeslotsScheduleById(
+				timeslotSchedule._id,
+				{ retrieveService: true, retrieveServiceProvider: true },
+			);
+			if (timeslotScheduleData._serviceProvider && !timeslotScheduleData._serviceProvider.service) {
+				timeslotScheduleData._serviceProvider.service = await this.servicesRepository.getService(
+					timeslotScheduleData._serviceProvider.serviceId,
+				);
+			}
+		}
+		const hasPermission = new TimeslotItemsActionAuthVisitor(timeslotScheduleData).hasPermission(authGroups);
+		if (!hasPermission) {
+			throw new MOLErrorV2(ErrorCodeV2.SYS_INVALID_AUTHORIZATION).setMessage(
+				`User cannot perform this timeslot item action for this service.`,
+			);
+		}
+	}
 
 	private static mapItemAndValidate(
 		timeslotsSchedule: TimeslotsSchedule,
@@ -58,6 +87,7 @@ export class TimeslotItemsService {
 			throw new MOLErrorV2(ErrorCodeV2.SYS_NOT_FOUND).setMessage('Timeslot item not found');
 		}
 
+		await this.verifyActionPermission(timeslotsSchedule);
 		return this.mapAndSaveTimeslotItem(timeslotsSchedule, request, timeslotItem);
 	}
 
@@ -65,10 +95,19 @@ export class TimeslotItemsService {
 		timeslotsSchedule: TimeslotsSchedule,
 		request: TimeslotItemRequest,
 	): Promise<TimeslotItem> {
+		await this.verifyActionPermission(timeslotsSchedule);
 		return this.mapAndSaveTimeslotItem(timeslotsSchedule, request, new TimeslotItem());
 	}
 
 	public async deleteTimeslot(timeslotId: number): Promise<DeleteResult> {
+		const timeslotSchedule = await this.getTimeslotsScheduleByTimeslotItemId(timeslotId);
+		await this.verifyActionPermission(timeslotSchedule);
+
 		return await this.timeslotItemsRepository.deleteTimeslotItem(timeslotId);
+	}
+
+	private async getTimeslotsScheduleByTimeslotItemId(id: number): Promise<TimeslotsSchedule> {
+		const timeslotItem = await this.timeslotItemsRepository.getTimeslotItem(id);
+		return this.timeslotsScheduleRepository.getTimeslotsScheduleById(timeslotItem._timeslotsScheduleId);
 	}
 }
